pyproject.toml
[tool.poetry]
name = "autoreelsmaker"
package-mode = false
version = "0.1.0"
description = ""
authors = ["Navicstein <navicsteinrotciv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
loguru = "^0.7.2"
moviepy = "^1.0.3"
aiohttp = "^3.9.5"
pydantic = "^2.8.2"
elevenlabs = "^1.3.1"
streamlit = "^1.36.0"
srt-equalizer = "^0.1.9"
termcolor = "^2.4.0"
opencv-python = "^4.10.0.84"
pytest-asyncio = "^0.23.8"
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
pytest-env = "^1.1.3"
langchain-core = "^0.2.28"
langchain = "^0.2.12"
langchain-openai = "^0.1.20"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.11"
pysrt = "^1.1.2"
word2number = "^1.1"
ffmpeg-python = "^0.2.0"
spacy = "^3.8.4"
pydub = "^0.25.1"
sqlmodel = "^0.0.22"
pydantic-settings = "^2.7.1"
together = "^1.3.14"
cuid = "^0.4"
cuid2 = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
